Added Login and JWT logic
-Added logic to set logged in status upon successful login and logout

-Added logic to check if logged in before allowing access to admin pages. Altough since this is client side its possible for users to override it but no sensitive information will be displayed if they would access admin pages as the JWT is required to get any information from the backend.

-Added jwt logic which gets set in a cookie for keeping track of in client and is then included in all api calls to backend.

-Added some error handling.

+++ Controllers/AdminController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Http.Json;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http.Json;
 using Microsoft.AspNetCore.Mvc;
 using RestaurantBookingFrontend.Models;
 using System.ComponentModel;
@@ -6,6 +7,8 @@ using System.Reflection;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using System.Web;
+using Microsoft.AspNetCore.Http;
 
 namespace RestaurantBookingFrontend.Controllers
 {
@@ -20,28 +23,134 @@ namespace RestaurantBookingFrontend.Controllers
             _httpClient = httpClient;
         }
 
+        //[Authorize]
         public IActionResult Index()
         {
+            // Return to login if not logged in
+            if(HttpContext.Session.GetString("IsLoggedIn") == "false")
+                return RedirectToAction("Login", "Admin");
+
+            var isLoggedIn = HttpContext.Session.GetString("IsLoggedIn");
+            ViewBag.IsLoggedIn = isLoggedIn;
+
             return View();
         }
 
+        [AllowAnonymous]
+        [HttpGet]
         public IActionResult Login()
         {
+            // Return to main admin page if logged in (no need to login twice)
+            if (HttpContext.Session.GetString("IsLoggedIn") == "true")
+                return RedirectToAction("Index", "Admin");
+
             return View();
         }
 
+        public IActionResult Logout()
+        {
+            // "Remove" jwt by making the contents an empty string and setting the expiration date in the past = not valid
+            Response.Cookies.Append("jwt", "", new CookieOptions
+            {
+                Expires = DateTime.Now.AddDays(-1),
+                HttpOnly = true,
+                Secure = true,
+                SameSite = SameSiteMode.Strict
+            });
 
+            // Set all logged in states to false
+            HttpContext.Session.SetString("IsLoggedIn", "false");
+            ViewBag.IsLoggedIn = "false";
+
+            // Set jwt in a cookie
+            var sessionCookieOptions = new CookieOptions
+            {
+                Path = "/",
+                Expires = DateTime.Now.AddDays(-1),
+                HttpOnly = true,
+                Secure = true,
+                SameSite = SameSiteMode.Strict
+            };
+            Response.Cookies.Append("JWT", "", sessionCookieOptions);
+
+            return RedirectToAction("Login", "Admin");
+        }
+
+
+        [AllowAnonymous]
+        [HttpPost("Admin/VerifyLogin")]
+        public async Task<IActionResult> VerifyLogin(LoginModel loginModel)
+        {
+            // Validate login
+            if (!ModelState.IsValid)
+                return View("Login", loginModel);
+                
+            // Create user object object
+            var user = new {username = loginModel.AdminUsername, password = loginModel.AdminPassword};
+
+            // Serialize
+            var json = JsonSerializer.Serialize(user);
+            var content = new StringContent(json, Encoding.UTF8, "application/json");
+
+            // Verify login
+            var response = await _httpClient.PostAsync($"{_baseUri}/api/User/AdminLogin", content);
+
+            if (!response.IsSuccessStatusCode)
+            {
+                ModelState.AddModelError("", $"Failed to login: {response.StatusCode} - {response.ReasonPhrase}");
+                return View("Login", loginModel);
+            }
+
+
+            // Deserialize json to jwt
+            var jsonString = await response.Content.ReadAsStreamAsync();
+            var jsonOptions = _jsonOptions;
+            var jwt = await JsonSerializer.DeserializeAsync<Jwt>(jsonString, jsonOptions);
+
+            if (jwt.Token == null)
+            {
+                ModelState.AddModelError("", $"Failed to login: An unexpected error occurred.");
+                return View("Login", loginModel);
+            }
+
+            // Set jwt in a cookie
+            var sessionCookieOptions = new CookieOptions
+            {
+                Path = "/",
+                Expires = DateTime.Now.AddHours(1),
+                HttpOnly = true,
+                Secure = true,
+                SameSite = SameSiteMode.Strict
+            };
+            Response.Cookies.Append("JWT", jwt.Token, sessionCookieOptions);
+
+
+            // Set logged in status to true to keep user logged in across the admin pages 
+            // Needed as frontend/backend is separated and we can't set [Authorize] on the pages - jwt only verifies access to endpoints
+            HttpContext.Session.SetString("IsLoggedIn", "true");
+
+            return RedirectToAction("Index", "Admin");
+        }
 
         public async Task<IActionResult> GetReservation(int reservationId)
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") == "false")
+                return RedirectToAction("Login", "Admin");
+
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
             var response = await _httpClient.GetAsync($"{_baseUri}/api/Reservation/GetReservationById?resId={reservationId}");
 
             if (!response.IsSuccessStatusCode)
-                return NotFound();
+                RedirectToAction("Login", "Admin");
 
             var jsonString = await response.Content.ReadAsStreamAsync();
 
             var jsonOptions = _jsonOptions;
+
 
             var reservation = await JsonSerializer.DeserializeAsync<ReservationById>(jsonString, jsonOptions);
 
@@ -50,10 +159,18 @@ namespace RestaurantBookingFrontend.Controllers
 
         public async Task<IActionResult> AllReservations()
         {
+            // Return to login if not logged in
+            //if (HttpContext.Session.GetString("IsLoggedIn") == "false")
+            //    return RedirectToAction("Login", "Admin");
+
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
             var response = await _httpClient.GetAsync($"{_baseUri}/api/Reservation/GetAllReservations");
 
             if (!response.IsSuccessStatusCode)
-                return NotFound();
+                return RedirectToAction("Login", "Admin");
 
             var jsonString = await response.Content.ReadAsStreamAsync();
             
@@ -65,8 +182,13 @@ namespace RestaurantBookingFrontend.Controllers
 
         [HttpPost("Admin/EditReservation")]
         public async Task<IActionResult> EditReservation(ReservationById reservationById)
-            {
-            if(!ModelState.IsValid)
+        {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") == "false")
+                return RedirectToAction("Login", "Admin");
+
+
+            if (!ModelState.IsValid)
                 return View(reservationById);
 
 
@@ -86,6 +208,10 @@ namespace RestaurantBookingFrontend.Controllers
 
             var content = new StringContent(json, Encoding.UTF8, "application/json");
 
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
             await _httpClient.PutAsync($"{_baseUri}/api/Reservation/UpdateReservation", content);
 
             return RedirectToAction("getReservation", new { reservationId = reservationById.Id });
@@ -95,10 +221,18 @@ namespace RestaurantBookingFrontend.Controllers
         [HttpGet]
         public async Task<IActionResult> EditReservation(int reservationId)
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") == "false")
+                return RedirectToAction("Login", "Admin");
+
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
             var response = await _httpClient.GetAsync($"{_baseUri}/api/Reservation/GetReservationById?resId={reservationId}");
 
             if (!response.IsSuccessStatusCode)
-                return NotFound();
+                RedirectToAction("Login", "Admin");
 
             var jsonString = await response.Content.ReadAsStreamAsync();
 
@@ -110,10 +244,18 @@ namespace RestaurantBookingFrontend.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteReservation(int reservationId)
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") != "true")
+                return RedirectToAction("Login", "Admin");
+
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
             var response = await _httpClient.DeleteAsync($"{_baseUri}/api/Reservation/DeleteReservation?resId={reservationId}");
 
             if (!response.IsSuccessStatusCode)
-                return NotFound();
+                RedirectToAction("Login", "Admin");
 
             return RedirectToAction("Index");
         }
@@ -121,10 +263,18 @@ namespace RestaurantBookingFrontend.Controllers
         [HttpGet("Admin/EditMenu")]
         public async Task<IActionResult> EditMenu()
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") != "true")
+                return RedirectToAction("Login", "Admin");
+
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
             var response = await _httpClient.GetAsync($"{_baseUri}/api/Dish/GetAllDishes");
 
             if (!response.IsSuccessStatusCode)
-                return NotFound();
+                RedirectToAction("Login", "Admin");
 
             var jsonString = await response.Content.ReadAsStreamAsync();
 
@@ -136,6 +286,14 @@ namespace RestaurantBookingFrontend.Controllers
         [HttpGet]
         public async Task<IActionResult> EditMenuItem(int dishId)
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") != "true")
+                return RedirectToAction("Login", "Admin");
+
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
             var response = await _httpClient.GetAsync($"{_baseUri}/api/Dish/GetDishById?dishId={dishId}");
 
             var jsonString = await response.Content.ReadAsStreamAsync();
@@ -151,11 +309,21 @@ namespace RestaurantBookingFrontend.Controllers
         [HttpPost]
         public async Task<IActionResult> UpdateMenuItem(Dish dish)
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") != "true")
+                return RedirectToAction("Login", "Admin");
+
             var json = JsonSerializer.Serialize(dish);
+
+            Console.WriteLine(json);
 
             var content = new StringContent(json, Encoding.UTF8, "application/json");
 
-            await _httpClient.PostAsync($"{_baseUri}/api/Dish/UpdateDish", content);
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
+            await _httpClient.PutAsync($"{_baseUri}/api/Dish/UpdateDish", content);
 
             return RedirectToAction("EditMenu");
         }
@@ -163,17 +331,28 @@ namespace RestaurantBookingFrontend.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteMenuItem(int dishId)
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") != "true")
+                return RedirectToAction("Login", "Admin");
+
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
+
             var response = await _httpClient.DeleteAsync($"{_baseUri}/api/Dish/DeleteDish?dishId={dishId}");
 
             if (!response.IsSuccessStatusCode)
-                return NotFound();
+                RedirectToAction("Login", "Admin");
 
             return RedirectToAction("EditMenu");
         }
 
-
         public async Task<IActionResult> AddMenuItem()
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") == "false")
+                return RedirectToAction("Login", "Admin");
+
             var dish = new Dish();
 
             return View(dish);
@@ -182,9 +361,17 @@ namespace RestaurantBookingFrontend.Controllers
         [HttpPost]
         public async Task<IActionResult> AddMenuItem(Dish dish)
         {
+            // Return to login if not logged in
+            if (HttpContext.Session.GetString("IsLoggedIn") != "true")
+                return RedirectToAction("Login", "Admin");
+
             var json = JsonSerializer.Serialize(dish);
 
             var content = new StringContent (json, Encoding.UTF8, "application/json");
+
+            // Get jwt from cookie and include it in the auth header for the api call
+            string? jwt = Request.Cookies["JWT"];
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", jwt);
 
             await _httpClient.PostAsync($"{_baseUri}/api/Dish/CreateDish", content);
